Finished User login and register functionalities

+++ base/views.py
@@ -1,4 +1,5 @@
 from multiprocessing import context
+from urllib import request
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from base.models import Room, Topic
@@ -7,6 +8,8 @@ from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages 
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 # Create your views here.
 
 '''
@@ -19,25 +22,46 @@ rooms = [
 
 def loginpage(request):
    
+   #if request.user.is_authenticated():
+    #  return redirect('home')
+   page = 'login'
+   context = {'page': page}
    if request.method == 'POST':
-      username = request.POST.get('username')
+      username = request.POST.get('username').lower()
       password = request.POST.get('password')
-   try:
-      user = User.objects.get(username=username)
-   except:
-      messages.error(request, 'User does not exist.')
+      try:
+         user = User.objects.get(username=username)
+      except:
+         messages.error(request, 'User does not exist.')
 
-   user = authenticate(request, username=username, password=password)
-   if user is not None:
-      login(request, user)
-      return redirect('home')
-   else:
-      messages.error(request, 'Username or Password not correct!')
-
-
-   return render(request, 'base/login_register.html')
+      user = authenticate(request, username=username, password=password)
+      if user is not None:
+         login(request, user)
+         return redirect('home')
+      else:
+         messages.error(request, 'Username or Password not correct!')
 
 
+   return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+
+   logout(request)
+   return redirect('home')
+def registerPage(request):
+
+   form = UserCreationForm()
+   if request.method == 'POST':
+      form = UserCreationForm(request.POST)
+      if form.is_valid():
+         user = form .save(commit=False)
+         user.username = user.username.lower()
+         user.save()
+         login(request, user)
+         return redirect('home')
+      else:
+         messages.error(request, 'Error occurred during registration')
+   return render(request, 'base/login_register.html', {'form': form})
 
 def room(request, pk):
    
@@ -64,7 +88,7 @@ def home(request):
 
    return  render(request, 'base/home.html',context)
 
-
+@login_required(login_url ='/login')
 def createRoom(request):
    form = RoomForm()
 
@@ -76,23 +100,27 @@ def createRoom(request):
       return redirect('home')
    return render(request, 'base/room_form.html', context)
 
+@login_required(login_url ='/login')
 def updateRoom(request, pk):
-   
-   
    room = Room.objects.get(id=pk) 
    form = RoomForm(instance=room)
    context = {'form': form}
-
+   if request.user != room.host:
+         return HttpResponse('You are not allowed here!!!')
    if request.method == 'POST':
       form = RoomForm(request.POST, instance=room)
+      
       if form.is_valid():
          form.save()
       return redirect('home')
    return render(request,'base/room_form.html', context)
 
+
+@login_required(login_url ='/login')
 def deleteRoom(request, pk):
    room = Room.objects.get(id=pk)
-   
+   if request.user != room.host:
+         return HttpResponse('You are not allowed here!!!')
    if request.method == 'POST':
       room.delete()
       return redirect('home')
